Creating separate source file for ArucoImage to reduce build dependency/time
